+++ src/components/combo-box/constants.js
@@ -1,3 +1,3 @@
-export const TOOLTIP_DELAY_MS = 60;
+export const TOOLTIP_DELAY_MS = 500;
 export const MAX_COUNT_OF_TOOLTIP_ITEMS = 10;
 export const SPINNER_DELAY = 50;