+++ webpack.docs.config.js
@@ -3,10 +3,13 @@ const webpack = require('webpack');
 const merge = require('webpack-merge');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const WebpackNotifierPlugin = require('webpack-notifier');
-const CleanWebpackPlugin = require('clean-webpack-plugin');
+const { CleanWebpackPlugin } = require('clean-webpack-plugin');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+const OptimizeCssAssetsPlugin = require('optimize-css-assets-webpack-plugin');
 const autoprefixer = require('autoprefixer');
 const precss = require('precss');
 const flexbugs = require('postcss-flexbugs-fixes');
+const packageConfig = require('./package.json');
 
 const isProd = process.env.NODE_ENV === 'production';
 
@@ -14,6 +17,9 @@ const PATHS = {
   build: path.join(__dirname, 'docs'),
   context: path.join(__dirname, 'src_docs'),
   jsFileName: 'examples.js',
+  jsChunkFileName: 'examples.[name].js',
+  cssFileName: 'examples.css',
+  cssChunkFileName: 'examples.[name].css',
   entry: path.join(__dirname, 'src_docs', 'index.jsx'),
   root: __dirname,
 };
@@ -24,12 +30,13 @@ const PATHS = {
 const baseConfig = {
   context: PATHS.context,
   entry: [
-    'babel-polyfill',
+    '@babel/polyfill',
     PATHS.entry,
   ],
   output: {
     path: PATHS.build,
     filename: PATHS.jsFileName,
+    chunkFilename: PATHS.jsChunkFileName,
   },
   module: {
     rules: [
@@ -47,59 +54,54 @@ const baseConfig = {
         ],
       },
       {
-        test: /\.css$/,
-        use: [
-          'style-loader',
-          'css-loader',
-          {
-            loader: 'postcss-loader',
-            options: {
-              plugins: () => [flexbugs, precss, autoprefixer],
-            },
-          },
-        ],
-      },
-      {
-        test: /\.scss$/,
-        use: [
-          'style-loader',
-          'css-loader',
-          {
-            loader: 'postcss-loader',
-            options: {
-              plugins: () => [flexbugs, precss, autoprefixer],
-            },
-          },
-          'sass-loader',
-        ],
-      },
-      {
         test: /\.svg$/,
-        use: ['babel-loader', 'react-svg-loader'],
+        use: [{
+          loader: 'babel-loader',
+        },
+        {
+          loader: 'react-svg-loader',
+        }],
       },
       {
         test: /\.(woff|woff2)(\?v=\d+\.\d+\.\d+)?$/,
-        use: [
-          'url-loader?limit=100&mimetype=application/font-woff&name=fonts/[name].[ext]',
-        ],
+        use: [{
+          loader: 'file-loader',
+          options: {
+            name: '[name].[ext]',
+            outputPath: 'fonts/',
+            mimetype: 'application/font-woff',
+          },
+        }],
       },
       {
         test: /\.ttf(\?v=\d+\.\d+\.\d+)?$/,
-        use: [
-          'url-loader?limit=100&mimetype=application/octet-stream&name=fonts/[name].[ext]',
-        ],
+        use: [{
+          loader: 'file-loader',
+          options: {
+            name: '[name].[ext]',
+            outputPath: 'fonts/',
+            mimetype: 'application/octet-stream',
+          },
+        }],
       },
       {
         test: /\.eot(\?v=\d+\.\d+\.\d+)?$/,
-        use: [
-          'file-loader?name=fonts/[name].[ext]',
-        ],
+        use: [{
+          loader: 'file-loader',
+          options: {
+            name: '[name].[ext]',
+            outputPath: 'fonts/',
+          },
+        }],
       },
       {
         test: /\.ico$/,
-        use: [
-          'file-loader?name=[name].[ext]',
-        ],
+        use: [{
+          loader: 'file-loader',
+          options: {
+            name: '[name].[ext]',
+          },
+        }],
       },
     ],
   },
@@ -107,10 +109,8 @@ const baseConfig = {
     fs: 'empty',
   },
   plugins: [
-    new CleanWebpackPlugin([PATHS.build], {
-      root: PATHS.root,
+    new CleanWebpackPlugin({
       verbose: true,
-      dry: false,
     }),
     new HtmlWebpackPlugin({
       filename: 'index.html',
@@ -123,6 +123,7 @@ const baseConfig = {
       'node_modules',
     ],
     extensions: ['.js', '.jsx'],
+    mainFields: ['es', 'cjs', 'browser', 'module', 'es:next', 'main'],
   },
 };
 
@@ -130,15 +131,36 @@ const baseConfig = {
 * DEVELOPMENT CONFIG
 */
 const devConfig = {
-  devtool: 'eval-source-map',
+  mode: 'development',
+  devtool: 'cheap-module-source-map',
+  module: {
+    rules: [
+      {
+        test: /\.(sa|sc|c)ss$/,
+        use: [
+          'style-loader',
+          'css-loader',
+          {
+            loader: 'postcss-loader',
+            options: {
+              plugins: () => [flexbugs, precss, autoprefixer],
+            },
+          },
+          'sass-loader',
+        ],
+      },
+    ],
+  },
   plugins: [
     new webpack.DefinePlugin({
       'process.env': {
         NODE_ENV: JSON.stringify('development'),
       },
     }),
-    new WebpackNotifierPlugin(),
-    new webpack.NamedModulesPlugin(),
+    new WebpackNotifierPlugin({
+      title: packageConfig.name,
+      contentImage: path.join(__dirname, 'src_docs', 'images', 'favicon.ico'),
+    }),
   ],
 };
 
@@ -146,25 +168,45 @@ const devConfig = {
 * PRODUCTION CONFIG
 */
 const prodConfig = {
-  devtool: 'source-map',
+  mode: 'production',
+  devtool: false,
+  module: {
+    rules: [
+      {
+        test: /\.(sa|sc|c)ss$/,
+        use: [
+          MiniCssExtractPlugin.loader,
+          'css-loader',
+          {
+            loader: 'postcss-loader',
+            options: {
+              plugins: () => [flexbugs, precss, autoprefixer],
+            },
+          },
+          'sass-loader',
+        ],
+      },
+    ],
+  },
   plugins: [
-    new webpack.optimize.CommonsChunkPlugin({
-      name: 'vendor',
-      filename: 'vendor.bundle.js',
-    }),
     new webpack.DefinePlugin({
       'process.env': {
         NODE_ENV: JSON.stringify('production'),
       },
     }),
-    new webpack.optimize.ModuleConcatenationPlugin(),
-    new webpack.optimize.UglifyJsPlugin({
-      sourceMap: true,
-      output: {
-        comments: false,
-      },
+    new MiniCssExtractPlugin({
+      filename: PATHS.cssFileName,
+      chunkFilename: PATHS.cssChunkFileName,
     }),
   ],
+  optimization: {
+    minimizer: [
+      new OptimizeCssAssetsPlugin({}),
+    ],
+    splitChunks: {
+      chunks: 'all',
+    },
+  },
 };
 
 module.exports = merge(baseConfig, isProd ? prodConfig : devConfig);